refactoring: overloaded signed bytes ctor of FileSize to eliminate code duplication

+++ logviewer.core/MainController.cs
@@ -667,7 +667,7 @@ namespace logviewer.core
         private void SetLogSize()
         {
             this.view.SetLoadedFileCapltion(this.currentPath);
-            this.view.HumanReadableLogSize = new FileSize((ulong)this.logSize, true).ToString();
+            this.view.HumanReadableLogSize = new FileSize(this.logSize, true).ToString();
         }
 
         public void ResetLogStatistic()
@@ -686,7 +686,7 @@ namespace logviewer.core
 
         public string GetLogSize(bool showBytes)
         {
-            return new FileSize((ulong)this.logSize, !showBytes).ToString();
+            return new FileSize(this.logSize, !showBytes).ToString();
         }
 
         public void ClearCache()
